Rename StorageRandomThings to StorageRandomGenerators for consistency
