+++ bag_verifier/src/test/scala/uk/ac/wellcome/platform/archive/bagverifier/generators/FixityGenerators.scala
@@ -7,11 +7,11 @@ import uk.ac.wellcome.platform.archive.bagverifier.fixity.{
   ExpectedFileFixity,
   FetchFileFixity
 }
-import uk.ac.wellcome.platform.archive.common.fixtures.StorageRandomThings
+import uk.ac.wellcome.platform.archive.common.generators.StorageRandomGenerators
 import uk.ac.wellcome.platform.archive.common.verify.{Checksum, MD5, SHA256}
 import uk.ac.wellcome.storage.Location
 
-trait FixityGenerators[BagLocation <: Location] extends StorageRandomThings {
+trait FixityGenerators[BagLocation <: Location] extends StorageRandomGenerators {
   def randomChecksum = Checksum(SHA256, randomChecksumValue)
   def badChecksum = Checksum(MD5, randomChecksumValue)
