+++ common/src/test/scala/uk/ac/wellcome/platform/archive/common/fixtures/OperationFixtures.scala
@@ -1,10 +1,11 @@
 package uk.ac.wellcome.platform.archive.common.fixtures
 
 import uk.ac.wellcome.messaging.memory.MemoryMessageSender
+import uk.ac.wellcome.platform.archive.common.generators.StorageRandomGenerators
 import uk.ac.wellcome.platform.archive.common.ingests.services.IngestUpdater
 import uk.ac.wellcome.platform.archive.common.operation.services.OutgoingPublisher
 
-trait OperationFixtures extends StorageRandomThings {
+trait OperationFixtures extends StorageRandomGenerators {
   def createIngestUpdaterWith(
     messageSender: MemoryMessageSender,
     stepName: String = createStepName
