+++ common/src/test/scala/uk/ac/wellcome/platform/archive/common/generators/PayloadOxumGenerators.scala
@@ -1,11 +1,10 @@
 package uk.ac.wellcome.platform.archive.common.generators
 
 import uk.ac.wellcome.platform.archive.common.bagit.models.PayloadOxum
-import uk.ac.wellcome.platform.archive.common.fixtures.StorageRandomThings
 
 import scala.util.Random
 
-trait PayloadOxumGenerators extends StorageRandomThings {
+trait PayloadOxumGenerators extends StorageRandomGenerators {
   def createPayloadOxumWith(
     payloadBytes: Long = Random.nextLong().abs,
     numberOfPayloadFiles: Int = randomInt(from = 1, to = 10000)
