+++ indexer/common/src/test/scala/uk/ac/wellcome/platform/archive/indexer/fixtures/IndexerFixtures.scala
@@ -7,14 +7,14 @@ import uk.ac.wellcome.akka.fixtures.Akka
 import uk.ac.wellcome.fixtures.TestWith
 import uk.ac.wellcome.messaging.fixtures.SQS.Queue
 import uk.ac.wellcome.messaging.fixtures.worker.AlpakkaSQSWorkerFixtures
-import uk.ac.wellcome.platform.archive.common.fixtures.StorageRandomThings
+import uk.ac.wellcome.platform.archive.common.generators.StorageRandomGenerators
 import uk.ac.wellcome.platform.archive.indexer.elasticsearch.IndexerWorker
 
 trait IndexerFixtures[SourceT, T, IndexedT]
     extends ElasticsearchFixtures
     with Akka
     with AlpakkaSQSWorkerFixtures
-    with StorageRandomThings { this: Suite =>
+    with StorageRandomGenerators { this: Suite =>
 
   def withIndexerWorker[R](
     index: Index,
